Restructure app to get ready to test Survey component
