+++ package.json
@@ -18,6 +18,7 @@
   },
   "devDependencies": {
     "babel-jest": "^5.3.0",
+    "fetch": "^0.3.6",
     "jest": "^0.1.40",
     "jest-cli": "^0.5.8",
     "jspm": "^0.16.10",
