+++ src/static/js/app/components/__tests__/SurveyAnswerGroup-test.js
@@ -1,13 +1,12 @@
 jest.dontMock('../SurveyAnswerGroup.js');
 
 describe('SurveyAnswerGroup', function () {
-  it('should start with nothing selected when questionAnswer is null', function () {
-    // TODO: Figure out how to reduce this duplication in every test:
-    var React = require('react');
-    var ReactDOM = require('react-dom');
-    var TestUtils = require('react-addons-test-utils');
-    var SurveyAnswerGroup = require('../SurveyAnswerGroup.js');
+  var React = require('react');
+  var ReactDOM = require('react-dom');
+  var TestUtils = require('react-addons-test-utils');
+  var SurveyAnswerGroup = require('../SurveyAnswerGroup.js');
 
+  it('should start with nothing selected when questionAnswer is null', function () {
     var answerGroup = TestUtils.renderIntoDocument(
       <SurveyAnswerGroup questionNumber={1} questionAnswer={null} />
     );
@@ -23,11 +22,6 @@ describe('SurveyAnswerGroup', function () {
   });
 
   it('should select the highest answer if questionAnswer is 5', function () {
-    var React = require('react');
-    var ReactDOM = require('react-dom');
-    var TestUtils = require('react-addons-test-utils');
-    var SurveyAnswerGroup = require('../SurveyAnswerGroup.js');
-
     var answerGroup = TestUtils.renderIntoDocument(
       <SurveyAnswerGroup questionNumber={1} questionAnswer={5} />
     );
@@ -36,11 +30,6 @@ describe('SurveyAnswerGroup', function () {
   });
 
   it('should select the middle answer if questionAnswer is 2', function () {
-    var React = require('react');
-    var ReactDOM = require('react-dom');
-    var TestUtils = require('react-addons-test-utils');
-    var SurveyAnswerGroup = require('../SurveyAnswerGroup.js');
-
     var answerGroup = TestUtils.renderIntoDocument(
       <SurveyAnswerGroup questionNumber={1} questionAnswer={2} />
     );
@@ -49,10 +38,6 @@ describe('SurveyAnswerGroup', function () {
   });
 
   it('should select the lowest answer if questionAnswer is 0', function () {
-    var React = require('react');
-    var TestUtils = require('react-addons-test-utils');
-    var SurveyAnswerGroup = require('../SurveyAnswerGroup.js');
-
     var answerGroup = TestUtils.renderIntoDocument(
       <SurveyAnswerGroup questionNumber={1} questionAnswer={0} />
     );
@@ -61,10 +46,6 @@ describe('SurveyAnswerGroup', function () {
   });
 
   it('should call a callback when an answer is selected with the answer provided', function () {
-    var React = require('react')
-    var TestUtils = require('react-addons-test-utils');
-    var SurveyAnswerGroup = require('../SurveyAnswerGroup.js');
-
     var dummyCallback = jest.genMockFunction();
 
     var answerGroup = TestUtils.renderIntoDocument(
